Mise à jour

+++ src/Services/DeleteItems.php
@@ -4,18 +4,28 @@ namespace App\Services;
 class DeleteItems
 {
     /**
-     * Suprime un dossier avec touts les fichiers inclus dedans
+     * Suprime un dossier avec tous les fichiers inclus dedans
      */
     public function deleteDirectoryAndHisFiles($path)
     {
+        //Initialisation de la variable pour supprimer le dossier
+        $eraseDir = true;
         if(is_dir($path)) {
             $objects = scandir($path);
+            //dd($objects);
             foreach($objects as $object) {
                 if($object != "." && $object != "..") {
-                    unlink($path . $object);
+                    if(is_file($path . $object)) {
+                        unlink($path . $object);
+                    }
+                    //Si le dossier contient d'autre dossier, par sécurité, on ne le supprime plus
+                    if(is_dir($path . $object)) {
+                        $eraseDir = false;
+                    }
+                    
                 }
             }
-            rmdir($path);
+            if($eraseDir) rmdir($path);
         }
     }
